graphics: Bug fix: Command splitting threshold is more strict on browsers?

+++ internal/graphics/image.go
@@ -94,7 +94,7 @@ func (i *Image) Fill(clr color.RGBA) error {
 	return nil
 }
 
-func (i *Image) DrawImage(src *Image, vertices []int16, geo, clr Matrix, mode opengl.CompositeMode) error {
+func (i *Image) DrawImage(src *Image, vertices []uint8, geo, clr Matrix, mode opengl.CompositeMode) error {
 	c := &drawImageCommand{
 		dst:      i,
 		src:      src,
